* develop: (127 commits)
  issue #3677: Wrong include fix.
  issue #3677: File rename. PerformanceEventDispatcher.h/.cpp -> PerformanceEventDispatcherTest.h/.cpp.
  issue #3677: Updates Windows project to add PerformanceEventDispatcherTest.
  issue #3677: Updates Android.mk and CMakeLists.txt.
  issue #3677: Adds EventDispatcher performance test.
  [AUTO] : updating submodule reference to latest autogenerated bindings
  Update CHANGELOG [ci skip]
  Update CHANGELOG [ci skip]
  [AUTO] : updating submodule reference to latest autogenerated bindings
  issue #3713: every autorelease will cause a release operation
  do not mix github access token with jenkins job trigger token
  issue #3781, Update travis-scripts, llvm-installing parts no need now.
  issue #3713: use reference in loop
  issue #3781, Minor fix & update submodule.
  closed #3808: EditBox on Mac works again.
  issue #3713:fix memory leak of autoreleas pool
  closed #3808: Removes unused files for MAC platform after using glfw3 to create opengl context
  global Z order uses floats
  ActionsProgress test include fixes
  Removed unused code and files.
  ...

Conflicts:
	cocos/2d/renderer/CCRenderer.cpp

+++ cocos/network/WebSocket.h
@@ -30,8 +30,11 @@
 #ifndef __CC_WEBSOCKET_H__
 #define __CC_WEBSOCKET_H__
 
-#include "cocos2d.h"
+#include "CCPlatformMacros.h"
+#include "CCStdC.h"
 #include <list>
+#include <string>
+#include <vector>
 
 struct libwebsocket;
 struct libwebsocket_context;
@@ -153,8 +156,8 @@ private:
     unsigned int _port;
     std::string  _path;
     
-    size_t _pendingFrameDataLen;
-    unsigned int _currentDataLen;
+    ssize_t _pendingFrameDataLen;
+    ssize_t _currentDataLen;
     char *_currentData;
 
     friend class WsThreadHelper;
